+++ reward_and_task_blueprint/templates/add_new_reward.html
